refactor: relocate components to route directories
