+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/page-header.tsx
@@ -17,10 +17,10 @@ import {
    useRef,
    useState,
 } from "react";
-import { ContextPanelHeaderActions } from "../routes/_authenticated/$slug/$teamSlug/-context-panel/context-panel-header-actions";
-import type { PanelAction } from "../routes/_authenticated/$slug/$teamSlug/-context-panel/context-panel-store";
-import { contextPanelStore } from "../routes/_authenticated/$slug/$teamSlug/-context-panel/context-panel-store";
-import { usePageActions } from "../routes/_authenticated/$slug/$teamSlug/-context-panel/use-context-panel";
+import { ContextPanelHeaderActions } from "../-context-panel/context-panel-header-actions";
+import type { PanelAction } from "../-context-panel/context-panel-store";
+import { contextPanelStore } from "../-context-panel/context-panel-store";
+import { usePageActions } from "../-context-panel/use-context-panel";
 
 export interface PageHeaderProps {
    title: string;
