+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/analytics/insights/-builder/insight-header.tsx
@@ -7,7 +7,7 @@ import {
    DropdownMenuTrigger,
 } from "@packages/ui/components/dropdown-menu";
 import { Copy, Ellipsis, Loader2, Save, Trash2 } from "lucide-react";
-import { PageHeader } from "@/components/page-header";
+import { PageHeader } from "../../../../-layout/page-header";
 
 interface InsightHeaderProps {
    name: string;
